+++ link_options.go
@@ -79,7 +79,7 @@ type SenderOptions struct {
 	// Default: ExpirySessionEnd.
 	TargetExpiryPolicy ExpiryPolicy
 
-	// ExpiryTimeout is the duration in seconds that the peer will be retained.
+	// TargetExpiryTimeout is the duration in seconds that the peer will be retained.
 	//
 	// Default: 0.
 	TargetExpiryTimeout uint32
@@ -180,25 +180,25 @@ type ReceiverOptions struct {
 	// TargetAddress specifies the target address for this receiver.
 	TargetAddress string
 
-	// SenderCapabilities is the list of extension capabilities the receiver desires.
-	SenderCapabilities []string
+	// SourceCapabilities is the list of extension capabilities the receiver desires.
+	SourceCapabilities []string
 
-	// SenderDurability indicates what state of the peer will be retained durably.
+	// SourceDurability indicates what state of the peer will be retained durably.
 	//
 	// Default: DurabilityNone.
-	SenderDurability Durability
+	SourceDurability Durability
 
-	// SenderExpiryPolicy determines when the expiry timer of the peer starts counting
+	// SourceExpiryPolicy determines when the expiry timer of the peer starts counting
 	// down from the timeout value.  If the link is subsequently re-attached before
 	// the timeout is reached, the count down is aborted.
 	//
 	// Default: ExpirySessionEnd.
-	SenderExpiryPolicy ExpiryPolicy
+	SourceExpiryPolicy ExpiryPolicy
 
-	// SenderExpiryTimeout is the duration in seconds that the peer will be retained.
+	// SourceExpiryTimeout is the duration in seconds that the peer will be retained.
 	//
 	// Default: 0.
-	SenderExpiryTimeout uint32
+	SourceExpiryTimeout uint32
 }
 
 // LinkFilter is an advanced API for setting non-standard source filters.